Let all Tasks extend "BuildToolsTask", giving better Utility methods. Remapping, Decompiling, Patching and preparing a build gradle does now work. Next step: Merging Bukkit / Spigot / PaperSpigot API and compile.

+++ src/main/java/de/heisluft/buildtools/tasks/ExtractMCSourcesTask.java
@@ -1,7 +1,6 @@
 package de.heisluft.buildtools.tasks;
 
 import de.heisluft.buildtools.utils.Utils;
-import org.gradle.api.DefaultTask;
 import org.gradle.api.tasks.TaskAction;
 
 import java.io.IOException;
@@ -13,7 +12,7 @@ import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.stream.Collectors;
 
-public class ExtractMCSourcesTask extends DefaultTask {
+public class ExtractMCSourcesTask extends BuildToolsTask {
 
   private static String rmTrailingSlash(Path p) {
     String s = p.toString();
@@ -22,8 +21,8 @@ public class ExtractMCSourcesTask extends DefaultTask {
 
   @TaskAction
   public void extractMCSources() {
-    Path base = Utils.<RemapServerJarTask>getTask(getProject(), "remapServerJar").getMappedJar()
-        .get().resolveSibling("decompiled").toAbsolutePath();
+    Path base = this.<RemapServerJarTask>getTask("remapServerJar").getMappedJar().get()
+        .resolveSibling("decompiled").toAbsolutePath();
     Path mJar = base.resolve("final-mapped.jar");
     try(FileSystem fs = FileSystems
         .newFileSystem(URI.create("jar:file:/" + mJar.toString().replace('\\', '/')),
@@ -37,7 +36,7 @@ public class ExtractMCSourcesTask extends DefaultTask {
         Path to = p.getFileName().toString().endsWith("java") ? srcMainJavaP
             .resolve(rmTrailingSlash(p)) : srcMainResP.resolve(rmTrailingSlash(p));
         if(!Files.isDirectory(to.getParent())) Files.createDirectories(to.getParent());
-        Files.copy(p, to);
+        if(!Files.exists(to)) Files.copy(p, to);
       }
     } catch(IOException e) {
       throw new RuntimeException(e);
